+++ fj-doc-base/src/main/java/org/fugerit/java/doc/base/typehelper/generic/SecurityHardeningConsts.java
@@ -1,0 +1,12 @@
+package org.fugerit.java.doc.base.typehelper.generic;
+
+public class SecurityHardeningConsts {
+
+    private SecurityHardeningConsts() {}
+
+    public static final int SECURITY_HARDENING_DISABLED = 0;
+
+    public static final int SECURITY_HARDENING_1 = 1;
+
+
+}